# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIMULA Backend",
    "description": "API PMR Mula. Router → service → SQLAlchemy ORM → PostgreSQL.",
    "version": "3.0.0"
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Me",
        "operationId": "me_api_v1_users_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "summary": "Profile Edit",
        "operationId": "profile_edit_api_v1_users_me_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileEdit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Staff Create",
        "operationId": "staff_create_api_v1_admin_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StaffCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/users/{uid}/active": {
      "patch": {
        "tags": [
          "Admin"
        ],
        "summary": "User State",
        "operationId": "user_state_api_v1_admin_users__uid__active_patch",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "uid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Uid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserState"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schools": {
      "get": {
        "tags": [
          "Schools"
        ],
        "summary": "Schools",
        "operationId": "schools_api_v1_schools_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SchoolOut"
                  },
                  "title": "Response Schools Api V1 Schools Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Schools"
        ],
        "summary": "School Create",
        "operationId": "school_create_api_v1_schools_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SchoolCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SchoolOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/modules/{mid}/quizzes": {
      "post": {
        "tags": [
          "Quizzes"
        ],
        "summary": "Quiz Create",
        "operationId": "quiz_create_api_v1_modules__mid__quizzes_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "mid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Mid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuizCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quizzes/{qid}": {
      "get": {
        "tags": [
          "Quizzes"
        ],
        "summary": "Quiz Detail",
        "operationId": "quiz_detail_api_v1_quizzes__qid__get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "qid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Qid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuizDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/quizzes/{qid}/attempts": {
      "post": {
        "tags": [
          "Quizzes"
        ],
        "summary": "Attempt Create",
        "operationId": "attempt_create_api_v1_quizzes__qid__attempts_post",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "qid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Qid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Submission"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/evaluation": {
      "get": {
        "tags": [
          "Reports"
        ],
        "summary": "Evaluation",
        "operationId": "evaluation_api_v1_reports_evaluation_get",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "school_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "exclusiveMinimum": 0
                },
                {
                  "type": "null"
                }
              ],
              "title": "School Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EvaluationOut"
                  },
                  "title": "Response Evaluation Api V1 Reports Evaluation Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EvaluationOut": {
        "properties": {
          "module_id": {
            "type": "integer",
            "title": "Module Id"
          },
          "participants": {
            "type": "integer",
            "title": "Participants"
          },
          "paired_participants": {
            "type": "integer",
            "title": "Paired Participants"
          },
          "mean_pre": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Pre"
          },
          "mean_post": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Post"
          },
          "mean_gain": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mean Gain"
          }
        },
        "type": "object",
        "required": [
          "module_id",
          "participants",
          "paired_participants",
          "mean_pre",
          "mean_post",
          "mean_gain"
        ],
        "title": "EvaluationOut"
      },
      "FeedbackOut": {
        "properties": {
          "question_id": {
            "type": "integer",
            "title": "Question Id"
          },
          "correct": {
            "type": "boolean",
            "title": "Correct"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "correct",
          "explanation"
        ],
        "title": "FeedbackOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileEdit": {
        "properties": {
          "display_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Display Name"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "locale": {
            "type": "string",
            "enum": [
              "id",
              "en"
            ],
            "title": "Locale",
            "default": "id"
          },
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "system"
            ],
            "title": "Theme",
            "default": "light"
          }
        },
        "type": "object",
        "required": [
          "display_name"
        ],
        "title": "ProfileEdit"
      },
...
      "QuestionCreate": {
        "properties": {
          "prompt": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Prompt"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 6,
            "minItems": 2,
            "title": "Options"
          },
          "correct_index": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Correct Index"
          },
          "explanation": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 3,
            "title": "Explanation"
          }
        },
        "type": "object",
        "required": [
          "prompt",
          "options",
          "correct_index",
          "explanation"
        ],
        "title": "QuestionCreate"
      },
...
      "QuestionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "options": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Options"
          }
        },
        "type": "object",
        "required": [
          "id",
          "prompt",
          "options"
        ],
        "title": "QuestionOut"
      },
      "QuizCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 3,
            "title": "Title"
          },
          "kind": {
            "type": "string",
            "enum": [
              "PRE",
              "POST"
            ],
            "title": "Kind"
          },
          "pass_score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Pass Score",
            "default": 70
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionCreate"
            },
            "type": "array",
            "maxItems": 50,
            "minItems": 1,
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "title",
          "kind",
          "questions"
        ],
        "title": "QuizCreate"
      },
      "QuizDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "module_id": {
            "type": "integer",
            "title": "Module Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "pass_score": {
            "type": "integer",
            "title": "Pass Score"
          },
          "questions": {
            "items": {
              "$ref": "#/components/schemas/QuestionOut"
            },
            "type": "array",
            "title": "Questions"
          }
        },
        "type": "object",
        "required": [
          "id",
          "module_id",
          "title",
          "kind",
          "pass_score",
          "questions"
        ],
        "title": "QuizDetail"
      },
      "QuizOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "module_id": {
            "type": "integer",
            "title": "Module Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "pass_score": {
            "type": "integer",
            "title": "Pass Score"
          }
        },
        "type": "object",
        "required": [
          "id",
          "module_id",
          "title",
          "kind",
          "pass_score"
        ],
        "title": "QuizOut"
      },
...
      "SchoolCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 2,
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "SchoolCreate"
      },
      "SchoolOut": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 150,
            "minLength": 2,
            "title": "Name"
          },
          "address": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Address"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id"
        ],
        "title": "SchoolOut"
      },
...
      "StaffCreate": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string",
                "format": "email"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "username": {
            "type": "string",
            "maxLength": 40,
            "minLength": 3,
            "pattern": "^[a-z0-9_]+$",
            "title": "Username"
          },
          "display_name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 2,
            "title": "Display Name"
          },
          "password": {
            "type": "string",
            "maxLength": 128,
            "minLength": 10,
            "title": "Password"
          },
          "school_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "School Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "STUDENT",
              "INSTRUCTOR",
              "REVIEWER",
              "ADMIN"
            ],
            "title": "Role",
            "default": "INSTRUCTOR"
          }
        },
        "type": "object",
        "required": [
          "username",
          "display_name",
          "password"
        ],
        "title": "StaffCreate"
      },
      "Submission": {
        "properties": {
          "answers": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "maxProperties": 50,
            "minProperties": 1,
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "Submission"
      },
      "SubmissionOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          },
          "quiz_id": {
            "type": "integer",
            "title": "Quiz Id"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "passed": {
            "type": "boolean",
            "title": "Passed"
          },
          "answers": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Answers"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "certificate_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Certificate Id"
          },
          "xp_awarded": {
            "type": "integer",
            "title": "Xp Awarded"
          },
          "feedback": {
            "items": {
              "$ref": "#/components/schemas/FeedbackOut"
            },
            "type": "array",
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "id",
          "user_id",
          "quiz_id",
          "score",
          "passed",
          "answers",
          "created_at",
          "xp_awarded",
          "feedback"
        ],
        "title": "SubmissionOut"
      },
...
      "UserOut": {
        "properties": {
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "locale": {
            "type": "string",
            "title": "Locale",
            "default": "id"
          },
          "theme": {
            "type": "string",
            "title": "Theme",
            "default": "light"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "username": {
            "type": "string",
            "title": "Username"
          },
          "display_name": {
            "type": "string",
            "title": "Display Name"
          },
          "role": {
            "type": "string",
            "title": "Role"
          },
          "school_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "School Id"
          },
          "xp": {
            "type": "integer",
            "title": "Xp"
          },
          "active": {
            "type": "boolean",
            "title": "Active",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "id",
          "username",
          "display_name",
          "role",
          "school_id",
          "xp"
        ],
        "title": "UserOut"
      },
      "UserState": {
        "properties": {
          "active": {
            "type": "boolean",
            "title": "Active"
          }
        },
        "type": "object",
        "required": [
          "active"
        ],
        "title": "UserState"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        }
      }
    }
  }
}